api/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
Django==3.2.5
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
django-extensions==3.1.3
django-rest-framework==0.1.0
djangorestframework==3.12.4
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
yapf==0.30.0
pillow==8.3.1
coreapi
